+++ Formula/jukeboxcli.rb
@@ -2,9 +2,8 @@ class Jukeboxcli < Formula
   desc "Mac-first terminal music player with local and online queues"
   homepage "https://github.com/funstuie-bit/JukeboxCli"
   url "https://github.com/funstuie-bit/JukeboxCli.git",
-      tag: "v0.1.1-beta.3",
-      revision: "14a28fa435904799b4926e4556ea67b321c6164c"
-  version "0.1.1-beta.3"
+      tag: "v0.1.1-beta.4"
+  version "0.1.1-beta.4"
   license "MIT"
 
   depends_on :macos